Update: Delete tour, change tour status
Update tour delete logic, set to disabled status instead if tour have booking
Add change tour status endpoint for change the tour status

+++ src/main/java/com/example/KDBS/controller/BookingController.java
@@ -119,7 +119,7 @@ public class BookingController {
     }
 
     @GetMapping("/company/{companyId}")
-    public ResponseEntity<BookingWithCountResponse> getBookingsByTourCompanyId(@PathVariable Long companyId) {
+    public ResponseEntity<BookingWithCountResponse> getBookingsByTourCompanyId(@PathVariable int companyId) {
         BookingWithCountResponse responses = bookingService.getAllBookingsByCompanyId(companyId);
         return ResponseEntity.ok(responses);
     }
